Add a systemd .service file
Some of the contrib files were changed to .m4 files, this is so we can
use the values set in the Makefile in them. This way, if you change the
bindir or the resdir when running make or make install, it'll be
reflected in the generated config file and systemd .service file.

Also, some of the default values in config.rs were changed to use env!()
calls to pull the values from the environment, which make supplies.

+++ Makefile
@@ -8,19 +8,12 @@
 
 SHELL		= /bin/sh
 
-INSTALLFLAGS	=
-INSTALL		= install $(INSTALLFLAGS)
-INSTALL_PROGRAM	= $(INSTALL)
-INSTALL_DATA	= $(INSTALL) -m 644
-
-CARGOFLAGS	= --locked
-CARGOFEATURES	= --all-features
-CARGO		= cargo $(CARGOFLAGS)
-
 prefix		= /usr/local
 exec_prefix	= $(prefix)
 
 bindir		= $(exec_prefix)/bin
+libdir		= $(exec_prefix)/lib
+servicedir	= $(libdir)/systemd/system
 datarootdir	= $(prefix)/share
 datadir		= $(datarootdir)
 mandir		= $(datarootdir)/man
@@ -35,42 +28,77 @@ localstatedir	= $(prefix)/var
 resdir		= $(localstatedir)/lib
 logdir		= $(localstatedir)/log
 
-target/release/longboard:
+INSTALLFLAGS	=
+INSTALL		= install $(INSTALLFLAGS)
+INSTALL_PROGRAM	= $(INSTALL)
+INSTALL_DATA	= $(INSTALL) -m 644
+
+CARGOFLAGS	= --locked
+CARGOFEATURES	= --all-features
+CARGO		= cargo $(CARGOFLAGS)
+
+M4FLAGS		=
+M4DEFINES	= -D BINDIR=$(bindir) -D RESDIR=$(resdir) -D LOGDIR=$(logdir)
+M4		= m4 $(M4DEFINES) $(M4FLAGS)
+
+export bindir
+export sysconfdir
+export resdir
+export logdir
+
+target/release/longboard: src/
 	$(CARGO) build $(CARGOFEATURES) --release
 
 all: target/release/longboard
 
 install: target/release/longboard
-	$(INSTALL_PROGRAM) -D target/release/longboard $(DESTDIR)$(bindir)/longboard
-	$(INSTALL_PROGRAM) -D target/release/longctl $(DESTDIR)$(bindir)/longctl
-	$(INSTALL_DATA) -D contrib/config/release.yaml \
-		$(DESTDIR)$(sysconfdir)/longboard/config.yaml
-	$(INSTALL_DATA) -D contrib/config/pages/rules.md \
-		$(DESTDIR)$(sysconfdir)/longboard/pages/rules.md
+	$(INSTALL_PROGRAM) -D target/release/longboard \
+		$(DESTDIR)$(bindir)/longboard
+	$(INSTALL_PROGRAM) -D target/release/longctl \
+		$(DESTDIR)$(bindir)/longctl
 	$(INSTALL_DATA) -D res/favicon.png -t $(DESTDIR)$(resdir)/longboard/
 	$(INSTALL_DATA) -D res/spoiler.png -t $(DESTDIR)$(resdir)/longboard/
-	$(INSTALL_DATA) -D res/banners/* -t $(DESTDIR)$(resdir)/longboard/banners
+	$(INSTALL_DATA) -D res/banners/* -t \
+		$(DESTDIR)$(resdir)/longboard/banners
 	$(INSTALL_DATA) -D res/script/* -t $(DESTDIR)$(resdir)/longboard/script
 	$(INSTALL_DATA) -D res/style/* -t $(DESTDIR)$(resdir)/longboard/style
 	cp -r res/templates $(DESTDIR)$(resdir)/longboard/templates
-	chown -R longboard:longboard $(DESTDIR)$(resdir)/longboard/templates
 	$(INSTALL_DATA) -d $(DESTDIR)$(resdir)/longboard/uploads
 	$(INSTALL_DATA) -d $(DESTDIR)$(logdir)/longboard
-	- mkdir -p $(DESTDIR)$(man1dir)
-	- mkdir -p $(DESTDIR)$(man5dir)
-	- mkdir -p $(DESTDIR)$(man8dir)
-	- gzip -c contrib/longctl.1 > $(DESTDIR)$(man1dir)/longctl$(man1ext).gz
-	- gzip -c contrib/longboard.5 > $(DESTDIR)$(man5dir)/longboard$(man5ext).gz
-	- gzip -c contrib/longboard.8 > $(DESTDIR)$(man8dir)/longboard$(man8ext).gz
+	mkdir -p $(DESTDIR)$(sysconfdir)/longboard
+	$(M4) contrib/config/release.yaml.m4 \
+		>$(DESTDIR)$(sysconfdir)/longboard/config.yaml
+	$(INSTALL_DATA) -D contrib/config/pages/rules.md \
+		$(DESTDIR)$(sysconfdir)/longboard/pages/rules.md
+ifdef servicedir
+	mkdir -p $(DESTDIR)$(servicedir)
+	$(M4) contrib/longboard.service.m4 \
+		>$(DESTDIR)$(servicedir)/longboard.service
+endif
+ifdef mandir
+	mkdir -p $(DESTDIR)$(man1dir)
+	mkdir -p $(DESTDIR)$(man5dir)
+	mkdir -p $(DESTDIR)$(man8dir)
+	gzip -c contrib/longctl.1 > $(DESTDIR)$(man1dir)/longctl$(man1ext).gz
+	gzip -c contrib/longboard.5 \
+		> $(DESTDIR)$(man5dir)/longboard$(man5ext).gz
+	gzip -c contrib/longboard.8 \
+		> $(DESTDIR)$(man8dir)/longboard$(man8ext).gz
+endif
 
 uninstall:
 	rm -f $(DESTDIR)$(bindir)/longboard
 	rm -f $(DESTDIR)$(bindir)/longctl
 	rm -rf $(DESTDIR)$(resdir)/longboard
 	rm -rf $(DESTDIR)$(logdir)/longboard
-	- rm -f $(DESTDIR)$(man5dir)/longboard$(man5ext).gz
-	- rm -f $(DESTDIR)$(man8dir)/longboard$(man8ext).gz
-	- rm -f $(DESTDIR)$(man1dir)/longctl$(man1ext).gz
+ifdef servicedir
+	rm -f $(DESTDIR)$(servicedir)/longboard.service
+endif
+ifdef mandir
+	rm -f $(DESTDIR)$(man5dir)/longboard$(man5ext).gz
+	rm -f $(DESTDIR)$(man8dir)/longboard$(man8ext).gz
+	rm -f $(DESTDIR)$(man1dir)/longctl$(man1ext).gz
+endif
 
 clean:
 	$(CARGO) clean
